feat: Serialize the empty string to undefined (#22)
BREAKING CHANGE: Serialization of the empty string is now treated as `undefined` instead of an UnserializableParamError.

Otherwise, the serialization of `{ foo: null }` would conflict with `{ foo: '' }`.
This serializer chooses to support the more common and more useful case of sending `null`
as a meaningful value while treating the empty string as sending no value.
This aligns with common UX patterns where input fields are initialized to the empty string.

Closes #20.